refactor: Encodable+Extension 파일 Requestable로 이동
