--- OpenMarket/OpenMarket/Extension/Encodable+Extension.swift
@@ -1,22 +1,0 @@
-//
-//  Encodable+Extension.swift
-//  OpenMarket
-//
-//  Created by Red, Mino on 2022/05/10.
-//
-
-import Foundation
-
-extension Encodable {
-    func toDictionary() -> Result<[String: Any], NetworkError> {
-        do {
-            let data = try JSONEncoder().encode(self)
-            guard let jsonData = try JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-                return .failure(.decodeError)
-            }
-            return .success(jsonData)
-        } catch {
-            return .failure(.decodeError)
-        }
-    }
-}
